+++ aiosmtpd/handlers.py
@@ -6,6 +6,7 @@ Pass in an instance of one of these classes, or derive your own, to provide
 your own handling of messages.  Implement only the methods you care about.
 """
 
+import re
 import sys
 import asyncio
 import logging
@@ -13,13 +14,13 @@ import mailbox
 import smtplib
 
 from email import message_from_bytes, message_from_string
-from email.feedparser import NLCRE
 from public import public
 
 
-EMPTYSTRING = ''
+EMPTYBYTES = b''
 COMMASPACE = ', '
-CRLF = '\r\n'
+CRLF = b'\r\n'
+NLCRE = re.compile(br'\r\n|\r|\n')
 log = logging.getLogger('mail.debug')
 
 
@@ -93,7 +94,11 @@ class Proxy:
 
     @asyncio.coroutine
     def handle_DATA(self, server, session, envelope):
-        lines = envelope.content.splitlines(keepends=True)
+        if isinstance(envelope.content, str):
+            content = envelope.original_content
+        else:
+            content = envelope.content
+        lines = content.splitlines(keepends=True)
         # Look for the last header
         i = 0
         ending = CRLF
@@ -102,8 +107,10 @@ class Proxy:
                 ending = line
                 break
             i += 1
-        lines.insert(i, 'X-Peer: %s%s' % (session.peer[0], ending))
-        data = EMPTYSTRING.join(lines)
+        peer = session.peer[0].encode('ascii')
+        # XXX Python 3.4 does not support %-interpolation for bytes objects.
+        lines.insert(i, b'X-Peer: ' + peer + ending)
+        data = EMPTYBYTES.join(lines)
         refused = self._deliver(envelope.mail_from, envelope.rcpt_tos, data)
         # TBD: what to do with refused addresses?
         log.info('we got some refusals: %s', refused)
@@ -122,7 +129,7 @@ class Proxy:
             log.info('got SMTPRecipientsRefused')
             refused = e.recipients
         except (OSError, smtplib.SMTPException) as e:
-            log.exception('got', e.__class__)
+            log.exception('got %s', e.__class__)
             # All recipients were refused.  If the exception had an associated
             # error code, use it.  Otherwise, fake it with a non-triggering
             # exception code.