+++ aiosmtpd/tests/test_smtp.py
@@ -7,17 +7,19 @@ import unittest
 from aiosmtpd.controller import Controller
 from aiosmtpd.handlers import Sink
 from aiosmtpd.smtp import SMTP as Server, __ident__ as GREETING
+from aiosmtpd.testing.helpers import reset_connection
 from contextlib import ExitStack
-from smtplib import SMTP, SMTPDataError, SMTPResponseException
-from unittest.mock import Mock, patch
+from smtplib import (
+    SMTP, SMTPDataError, SMTPResponseException, SMTPServerDisconnected)
+from unittest.mock import Mock, PropertyMock, patch
 
 CRLF = '\r\n'
 BCRLF = b'\r\n'
 
 
-class UTF8Controller(Controller):
+class DecodingController(Controller):
     def factory(self):
-        return Server(self.handler, decode_data=True)
+        return Server(self.handler, decode_data=True, enable_SMTPUTF8=True)
 
 
 class NoDecodeController(Controller):
@@ -33,7 +35,7 @@ class ReceivingHandler:
 
     @asyncio.coroutine
     def handle_DATA(self, server, session, envelope):
-        self.box.append(envelope.content)
+        self.box.append(envelope)
         return '250 OK'
 
 
@@ -46,9 +48,9 @@ class SizedController(Controller):
         return Server(self.handler, data_size_limit=self.size)
 
 
-class SMTPUTF8Controller(Controller):
+class StrictASCIIController(Controller):
     def factory(self):
-        return Server(self.handler, enable_SMTPUTF8=True)
+        return Server(self.handler, enable_SMTPUTF8=False, decode_data=True)
 
 
 class CustomHostnameController(Controller):
@@ -73,6 +75,40 @@ class ErroringHandler:
     @asyncio.coroutine
     def handle_exception(self, error):
         self.error = error
+        return '500 ErroringHandler handling error'
+
+
+class ErroringHandlerCustomResponse:
+    error = None
+
+    @asyncio.coroutine
+    def handle_exception(self, error):
+        self.error = error
+        return '451 Temporary error: ({}) {}'.format(
+            error.__class__.__name__, str(error))
+
+
+class ErroringErrorHandler:
+    error = None
+
+    @asyncio.coroutine
+    def handle_exception(self, error):
+        self.error = error
+        raise ValueError('ErroringErrorHandler test')
+
+
+class UndescribableError(Exception):
+    def __str__(self):
+        raise Exception()
+
+
+class UndescribableErrorHandler:
+    error = None
+
+    @asyncio.coroutine
+    def handle_exception(self, error):
+        self.error = error
+        raise UndescribableError()
 
 
 class ErrorSMTP(Server):
@@ -124,7 +160,7 @@ class TestProtocol(unittest.TestCase):
         except asyncio.CancelledError:
             pass
         self.assertEqual(len(handler.box), 1)
-        self.assertEqual(handler.box[0], data)
+        self.assertEqual(handler.box[0].content, data)
 
     def test_empty_email(self):
         handler = ReceivingHandler()
@@ -143,12 +179,12 @@ class TestProtocol(unittest.TestCase):
             pass
         self.assertEqual(self.responses[5], b'250 OK\r\n')
         self.assertEqual(len(handler.box), 1)
-        self.assertEqual(handler.box[0], b'')
+        self.assertEqual(handler.box[0].content, b'')
 
 
 class TestSMTP(unittest.TestCase):
     def setUp(self):
-        controller = UTF8Controller(Sink)
+        controller = DecodingController(Sink)
         controller.start()
         self.addCleanup(controller.stop)
         self.address = (controller.hostname, controller.port)
@@ -182,7 +218,8 @@ class TestSMTP(unittest.TestCase):
             lines = response.splitlines()
             self.assertEqual(lines[0], bytes(socket.getfqdn(), 'utf-8'))
             self.assertEqual(lines[1], b'SIZE 33554432')
-            self.assertEqual(lines[2], b'HELP')
+            self.assertEqual(lines[2], b'SMTPUTF8')
+            self.assertEqual(lines[3], b'HELP')
 
     def test_ehlo_duplicate(self):
         with SMTP(*self.address) as client:
@@ -434,6 +471,16 @@ class TestSMTP(unittest.TestCase):
                 response,
                 b'Syntax: MAIL FROM: <address> [SP <mail-parameters>]')
 
+    # Test the workaround http://bugs.python.org/issue27931
+    @patch('email._header_value_parser.AngleAddr.addr_spec',
+           side_effect=IndexError, new_callable=PropertyMock)
+    def test_mail_fail_parse_email(self, addr_spec):
+        with SMTP(*self.address) as client:
+            client.helo('example.com')
+            code, response = client.docmd('MAIL FROM: <""@example.com>')
+            self.assertEqual(code, 501)
+            self.assertEqual(response, b'Syntax: MAIL FROM: <address>')
+
     def test_rcpt_no_helo(self):
         with SMTP(*self.address) as client:
             code, response = client.docmd('RCPT TO: <anne@example.com>')
@@ -455,6 +502,16 @@ class TestSMTP(unittest.TestCase):
             code, response = client.docmd('MAIL FROM: <anne@example.com>')
             self.assertEqual(code, 250)
             code, response = client.docmd('RCPT')
+            self.assertEqual(code, 501)
+            self.assertEqual(response, b'Syntax: RCPT TO: <address>')
+
+    def test_rcpt_no_to(self):
+        with SMTP(*self.address) as client:
+            code, response = client.helo('example.com')
+            self.assertEqual(code, 250)
+            code, response = client.docmd('MAIL FROM: <anne@example.com>')
+            self.assertEqual(code, 250)
+            code, response = client.docmd('RCPT <anne@example.com')
             self.assertEqual(code, 501)
             self.assertEqual(response, b'Syntax: RCPT TO: <address>')
 
@@ -518,6 +575,22 @@ class TestSMTP(unittest.TestCase):
             self.assertEqual(
                 response,
                 b'RCPT TO parameters not recognized or not implemented')
+
+    # Test the workaround http://bugs.python.org/issue27931
+    @patch('email._header_value_parser.AngleAddr.addr_spec',
+           new_callable=PropertyMock)
+    def test_rcpt_fail_parse_email(self, addr_spec):
+        with SMTP(*self.address) as client:
+            code, response = client.ehlo('example.com')
+            self.assertEqual(code, 250)
+            code, response = client.docmd('MAIL FROM: <anne@example.com>')
+            self.assertEqual(code, 250)
+            addr_spec.side_effect = IndexError
+            code, response = client.docmd('RCPT TO: <""@example.com>')
+            self.assertEqual(code, 501)
+            self.assertEqual(
+                response,
+                b'Syntax: RCPT TO: <address> [SP <mail-parameters>]')
 
     def test_rset(self):
         with SMTP(*self.address) as client:
@@ -614,18 +687,34 @@ class TestSMTPWithController(unittest.TestCase):
                 b'Error: message size exceeds fixed maximum message size')
 
     def test_mail_with_compatible_smtputf8(self):
-        controller = SMTPUTF8Controller(Sink())
+        handler = ReceivingHandler()
+        controller = Controller(handler)
         controller.start()
         self.addCleanup(controller.stop)
+        recipient = 'bart\xCB@example.com'
+        sender = 'anne\xCB@example.com'
         with SMTP(controller.hostname, controller.port) as client:
             client.ehlo('example.com')
-            code, response = client.docmd(
-                'MAIL FROM: <anne@example.com> SMTPUTF8')
+            client.send(bytes(
+                'MAIL FROM: <' + sender + '> SMTPUTF8\r\n',
+                encoding='utf-8'))
+            code, response = client.getreply()
             self.assertEqual(code, 250)
             self.assertEqual(response, b'OK')
+            client.send(bytes(
+                'RCPT TO: <' + recipient + '>\r\n',
+                encoding='utf-8'))
+            code, response = client.getreply()
+            self.assertEqual(code, 250)
+            self.assertEqual(response, b'OK')
+            code, response = client.data('')
+            self.assertEqual(code, 250)
+            self.assertEqual(response, b'OK')
+        self.assertEqual(handler.box[0].rcpt_tos[0], recipient)
+        self.assertEqual(handler.box[0].mail_from, sender)
 
     def test_mail_with_unrequited_smtputf8(self):
-        controller = SMTPUTF8Controller(Sink())
+        controller = Controller(Sink())
         controller.start()
         self.addCleanup(controller.stop)
         with SMTP(controller.hostname, controller.port) as client:
@@ -635,7 +724,7 @@ class TestSMTPWithController(unittest.TestCase):
             self.assertEqual(response, b'OK')
 
     def test_mail_with_incompatible_smtputf8(self):
-        controller = SMTPUTF8Controller(Sink())
+        controller = Controller(Sink())
         controller.start()
         self.addCleanup(controller.stop)
         with SMTP(controller.hostname, controller.port) as client:
@@ -701,7 +790,7 @@ Testing
 
     def test_dots_escaped(self):
         handler = ReceivingHandler()
-        controller = UTF8Controller(handler)
+        controller = DecodingController(handler)
         controller.start()
         self.addCleanup(controller.stop)
         with SMTP(controller.hostname, controller.port) as client:
@@ -709,7 +798,7 @@ Testing
             mail = CRLF.join(['Test', '.', 'mail'])
             client.sendmail('anne@example.com', ['bart@example.com'], mail)
             self.assertEqual(len(handler.box), 1)
-            self.assertEqual(handler.box[0], 'Test\r\n.\r\nmail')
+            self.assertEqual(handler.box[0].content, 'Test\r\n.\r\nmail')
 
     def test_unexpected_errors(self):
         handler = ErroringHandler()
@@ -724,7 +813,7 @@ Testing
                 SMTP(controller.hostname, controller.port))
             code, response = client.helo('example.com')
         self.assertEqual(code, 500)
-        self.assertEqual(response, b'Error: (ValueError) test')
+        self.assertEqual(response, b'ErroringHandler handling error')
         self.assertIsInstance(handler.error, ValueError)
 
     def test_unexpected_errors_unhandled(self):
@@ -745,6 +834,81 @@ Testing
         # handler.error did not change because the handler does not have a
         # handle_exception() method.
         self.assertIsNone(handler.error)
+
+    def test_unexpected_errors_custom_response(self):
+        handler = ErroringHandlerCustomResponse()
+        controller = ErrorController(handler)
+        controller.start()
+        self.addCleanup(controller.stop)
+        with ExitStack() as resources:
+            # Suppress logging to the console during the tests.  Depending on
+            # timing, the exception may or may not be logged.
+            resources.enter_context(patch('aiosmtpd.smtp.log.exception'))
+            client = resources.enter_context(
+                SMTP(controller.hostname, controller.port))
+            code, response = client.helo('example.com')
+        self.assertEqual(code, 451)
+        self.assertEqual(response, b'Temporary error: (ValueError) test')
+        self.assertIsInstance(handler.error, ValueError)
+
+    def test_exception_handler_exception(self):
+        handler = ErroringErrorHandler()
+        controller = ErrorController(handler)
+        controller.start()
+        self.addCleanup(controller.stop)
+        with ExitStack() as resources:
+            # Suppress logging to the console during the tests.  Depending on
+            # timing, the exception may or may not be logged.
+            resources.enter_context(patch('aiosmtpd.smtp.log.exception'))
+            client = resources.enter_context(
+                SMTP(controller.hostname, controller.port))
+            code, response = client.helo('example.com')
+        self.assertEqual(code, 500)
+        self.assertEqual(response,
+                         b'Error: (ValueError) ErroringErrorHandler test')
+        self.assertIsInstance(handler.error, ValueError)
+
+    def test_exception_handler_undescribable(self):
+        handler = UndescribableErrorHandler()
+        controller = ErrorController(handler)
+        controller.start()
+        self.addCleanup(controller.stop)
+        with ExitStack() as resources:
+            # Suppress logging to the console during the tests.  Depending on
+            # timing, the exception may or may not be logged.
+            resources.enter_context(patch('aiosmtpd.smtp.log.exception'))
+            client = resources.enter_context(
+                SMTP(controller.hostname, controller.port))
+            code, response = client.helo('example.com')
+        self.assertEqual(code, 500)
+        self.assertEqual(response, b'Error: Cannot describe error')
+        self.assertIsInstance(handler.error, ValueError)
+
+    def test_bad_encodings(self):
+        handler = ReceivingHandler()
+        controller = DecodingController(handler)
+        controller.start()
+        self.addCleanup(controller.stop)
+        with SMTP(controller.hostname, controller.port) as client:
+            client.helo('example.com')
+            mail_from = b'anne\xFF@example.com'
+            mail_to = b'bart\xFF@example.com'
+            client.ehlo('test')
+            client.send(b'MAIL FROM:' + mail_from + b'\r\n')
+            code, response = client.getreply()
+            self.assertEqual(code, 250)
+            client.send(b'RCPT TO:' + mail_to + b'\r\n')
+            code, response = client.getreply()
+            self.assertEqual(code, 250)
+            client.data('Test mail')
+            self.assertEqual(len(handler.box), 1)
+            envelope = handler.box[0]
+            mail_from2 = envelope.mail_from.encode(
+                'utf-8', errors='surrogateescape')
+            self.assertEqual(mail_from2, mail_from)
+            mail_to2 = envelope.rcpt_tos[0].encode(
+                'utf-8', errors='surrogateescape')
+            self.assertEqual(mail_to2, mail_to)
 
 
 class TestCustomizations(unittest.TestCase):
@@ -790,3 +954,106 @@ class TestCustomizations(unittest.TestCase):
             self.assertEqual(
                 response,
                 b'Error: BODY can only be one of 7BIT, 8BITMIME')
+
+
+class TestClientCrash(unittest.TestCase):
+    # GH#62 - if the client crashes during the SMTP dialog we want to make
+    # sure we don't get tracebacks where we call readline().
+    def setUp(self):
+        controller = Controller(Sink)
+        controller.start()
+        self.addCleanup(controller.stop)
+        self.address = (controller.hostname, controller.port)
+
+    def test_connection_reset_during_DATA(self):
+        with SMTP(*self.address) as client:
+            client.helo('example.com')
+            client.docmd('MAIL FROM: <anne@example.com>')
+            client.docmd('RCPT TO: <bart@example.com>')
+            client.docmd('DATA')
+            # Start sending the DATA but reset the connection before that
+            # completes, i.e. before the .\r\n
+            client.send(b'From: <anne@example.com>')
+            reset_connection(client)
+            # The connection should be disconnected, so trying to do another
+            # command from here will give us an exception.  In GH#62, the
+            # server just hung.
+            self.assertRaises(SMTPServerDisconnected, client.noop)
+
+    def test_connection_reset_during_command(self):
+        with SMTP(*self.address) as client:
+            client.helo('example.com')
+            # Start sending a command but reset the connection before that
+            # completes, i.e. before the \r\n
+            client.send('MAIL FROM: <anne')
+            reset_connection(client)
+            # The connection should be disconnected, so trying to do another
+            # command from here will give us an exception.  In GH#62, the
+            # server just hung.
+            self.assertRaises(SMTPServerDisconnected, client.noop)
+
+    def test_close_in_command(self):
+        with SMTP(*self.address) as client:
+            # Don't include the CRLF.
+            client.send('FOO')
+            client.close()
+
+    def test_close_in_data(self):
+        with SMTP(*self.address) as client:
+            code, response = client.helo('example.com')
+            self.assertEqual(code, 250)
+            code, response = client.docmd('MAIL FROM: <anne@example.com>')
+            self.assertEqual(code, 250)
+            code, response = client.docmd('RCPT TO: <bart@example.com>')
+            self.assertEqual(code, 250)
+            code, response = client.docmd('DATA')
+            self.assertEqual(code, 354)
+            # Don't include the CRLF.
+            client.send('FOO')
+            client.close()
+
+
+class TestStrictASCII(unittest.TestCase):
+    def setUp(self):
+        controller = StrictASCIIController(Sink())
+        controller.start()
+        self.addCleanup(controller.stop)
+        self.address = (controller.hostname, controller.port)
+
+    def test_ehlo(self):
+        with SMTP(*self.address) as client:
+            code, response = client.ehlo('example.com')
+            self.assertEqual(code, 250)
+            lines = response.splitlines()
+            self.assertNotIn(b'SMTPUTF8', lines)
+
+    def test_bad_encoded_param(self):
+        with SMTP(*self.address) as client:
+            client.ehlo('example.com')
+            client.send(b'MAIL FROM: <anne\xFF@example.com>\r\n')
+            code, response = client.getreply()
+            self.assertEqual(code, 500)
+            self.assertIn(b'Error: strict ASCII mode', response)
+
+    def test_mail_param(self):
+        with SMTP(*self.address) as client:
+            client.ehlo('example.com')
+            code, response = client.docmd(
+                'MAIL FROM: <anne@example.com> SMTPUTF8')
+            self.assertEqual(code, 501)
+            self.assertEqual(response, b'Error: SMTPUTF8 disabled')
+
+    def test_data(self):
+        with SMTP(*self.address) as client:
+            code, response = client.ehlo('example.com')
+            self.assertEqual(code, 250)
+            with self.assertRaises(SMTPDataError) as cm:
+                client.sendmail('anne@example.com', ['bart@example.com'], b"""\
+From: anne@example.com
+To: bart@example.com
+Subject: A test
+
+Testing\xFF
+""")
+            self.assertEqual(cm.exception.smtp_code, 500)
+            self.assertIn(b'Error: strict ASCII mode', cm.exception.smtp_error)